G main window refactor (#1)
* Started GMainWindow refactor

* Finished refactor

* Remove unintended whitespace diff

* Removed more unintended diffs

* Removing unneeded headers from realmain.cpp

* Removed more needless headers from realmain.cpp

* Finished removing needless headers

* Whitespace delta

* Working build

* Github needs to specify
