+++ src/citra_qt/CMakeLists.txt
@@ -8,6 +8,9 @@ if (POLICY CMP0071)
 endif()
 
 add_executable(citra-qt
+    realmain.cpp
+)
+add_library(citra-qt-lib STATIC
     Info.plist
     aboutdialog.cpp
     aboutdialog.h
@@ -247,9 +250,13 @@ endif()
 
 create_target_directory_groups(citra-qt)
 
-target_link_libraries(citra-qt PRIVATE audio_core common core input_common network video_core vanguard vanguardwrapper)
-target_link_libraries(citra-qt PRIVATE Boost::boost glad nihstro-headers Qt5::OpenGL Qt5::Widgets Qt5::Multimedia)
-target_link_libraries(citra-qt PRIVATE ${PLATFORM_LIBRARIES} Threads::Threads)
+target_include_directories(citra-qt PRIVATE $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}/citra-qt-lib_autogen/include_${CMAKE_BUILD_TYPE}>)
+target_link_libraries(citra-qt PRIVATE citra-qt-lib vanguardwrapper)
+target_link_libraries(citra-qt PRIVATE ${PLATFORM_LIBRARIES} nihstro-headers Qt5::Multimedia)
+
+target_link_libraries(citra-qt-lib PRIVATE audio_core common core input_common network video_core vanguard)
+target_link_libraries(citra-qt-lib PRIVATE Boost::boost glad nihstro-headers Qt5::OpenGL Qt5::Widgets Qt5::Multimedia)
+target_link_libraries(citra-qt-lib PRIVATE Threads::Threads)
 
 target_compile_definitions(citra-qt PRIVATE
     # Use QStringBuilder for string concatenation to reduce
