+++ src/citra_qt/main.cpp
@@ -99,13 +99,6 @@
 Q_IMPORT_PLUGIN(QWindowsIntegrationPlugin);
 #endif
 
-#ifdef _WIN32
-extern "C" {
-// tells Nvidia drivers to use the dedicated GPU by default on laptops with switchable graphics
-__declspec(dllexport) unsigned long NvOptimusEnablement = 0x00000001;
-}
-#endif
-
 constexpr int default_mouse_timeout = 2500;
 
 /**
@@ -2410,62 +2403,4 @@ void GMainWindow::SetDiscordEnabled([[maybe_unused]] bool state) {
     discord_rpc = std::make_unique<DiscordRPC::NullImpl>();
 #endif
     discord_rpc->Update();
-}
-
-#ifdef main
-#undef main
-#endif
-
-int main(int argc, char* argv[]) {
-    Common::DetachedTasks detached_tasks;
-    MicroProfileOnThreadCreate("Frontend");
-    SCOPE_EXIT({ MicroProfileShutdown(); });
-
-    // Init settings params
-    QCoreApplication::setOrganizationName(QStringLiteral("Citra team"));
-    QCoreApplication::setApplicationName(QStringLiteral("Citra"));
-
-    QSurfaceFormat format;
-    format.setVersion(3, 3);
-    format.setProfile(QSurfaceFormat::CoreProfile);
-    format.setSwapInterval(0);
-    // TODO: expose a setting for buffer value (ie default/single/double/triple)
-    format.setSwapBehavior(QSurfaceFormat::DefaultSwapBehavior);
-    QSurfaceFormat::setDefaultFormat(format);
-
-#ifdef __APPLE__
-    std::string bin_path = FileUtil::GetBundleDirectory() + DIR_SEP + "..";
-    chdir(bin_path.c_str());
-#endif
-    QCoreApplication::setAttribute(Qt::AA_DontCheckOpenGLContextThreadAffinity);
-    QCoreApplication::setAttribute(Qt::AA_ShareOpenGLContexts);
-    QApplication app(argc, argv);
-
-    // Qt changes the locale and causes issues in float conversion using std::to_string() when
-    // generating shaders
-    setlocale(LC_ALL, "C");
-
-    GMainWindow main_window;
-
-    // Register CameraFactory
-    Camera::RegisterFactory("image", std::make_unique<Camera::StillImageCameraFactory>());
-    Camera::RegisterFactory("qt", std::make_unique<Camera::QtMultimediaCameraFactory>());
-    Camera::QtMultimediaCameraHandler::Init();
-
-    // Register frontend applets
-    Frontend::RegisterDefaultApplets();
-    Core::System::GetInstance().RegisterMiiSelector(std::make_shared<QtMiiSelector>(main_window));
-    Core::System::GetInstance().RegisterSoftwareKeyboard(std::make_shared<QtKeyboard>(main_window));
-
-    // Register Qt image interface
-    Core::System::GetInstance().RegisterImageInterface(std::make_shared<QtImageInterface>());
-
-    main_window.show();
-
-    QObject::connect(&app, &QGuiApplication::applicationStateChanged, &main_window,
-                     &GMainWindow::OnAppFocusStateChanged);
-
-    int result = app.exec();
-    detached_tasks.WaitForAllTasks();
-    return result;
 }
